VideoRenderers: вместо CMap используем std::unordered_map, упрощение избыточного кода, косметика.
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@3262 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/filters/renderer/VideoRenderers/SyncRenderer.h
@@ -1,6 +1,6 @@
 /*
  * (C) 2003-2006 Gabest
- * (C) 2006-2017 see Authors.txt
+ * (C) 2006-2018 see Authors.txt
  *
  * This file is part of MPC-BE.
  *
@@ -233,7 +233,7 @@ namespace GothSync
 			_Out_ LPD3DXSPRITE      *ppSprite
 		);
 
-		int m_nTearingPos;
+		long m_nTearingPos;
 		VMR9AlphaBitmap m_VMR9AlphaBitmap;
 		CAutoVectorPtr<BYTE> m_VMR9AlphaBitmapData;
 		CRect m_VMR9AlphaBitmapRect;
